Move services to separate module in opentherm_gw
